refactor: extract translucent-screen compositing module
scene_radiance.py had grown past the 1000-line module cap.  The
translucent-screen ordering (ScreenOp, translucent_screen_ops) and the
depth-ambiguity checks move to the sibling scene_compositing.py, with
cross-referencing docstrings, an api-reference entry, and a dev-guide
module-table row; the radiance stage paints the stack and applies the
ordered ops.  Pure code motion: the full scene catalog renders
bit-identically.
