+++ docs/api_reference/api_sim.rst
@@ -28,6 +28,11 @@ spindoctor.sim
    :undoc-members:
    :show-inheritance:
 
+.. automodule:: spindoctor.sim.forward.scene_compositing
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 .. automodule:: spindoctor.sim.forward.body
    :members:
    :undoc-members:
